+++ src/schemathesis/generation/hypothesis/builder.py
@@ -21,7 +21,6 @@ from hypothesis import Phase, Verbosity
 from hypothesis import strategies as st
 from hypothesis._settings import all_settings
 from hypothesis.errors import Unsatisfiable
-from jsonschema.exceptions import SchemaError
 from jsonschema_rs import ValidationError
 from requests.models import CaseInsensitiveDict
 
@@ -350,7 +349,6 @@ def generate_example_cases(
         Unsatisfiable,
         UnresolvableReference,
         SerializationNotPossible,
-        SchemaError,
         ValidationError,
     ) as exc:
         result = []
@@ -358,8 +356,6 @@ def generate_example_cases(
             UnsatisfiableExampleMark.set(test, exc)
         if isinstance(exc, SerializationNotPossible):
             NonSerializableMark.set(test, exc)
-        if isinstance(exc, SchemaError):
-            InvalidRegexMark.set(test, exc)
         if is_regex_validation_error(exc):
             InvalidRegexMark.set(test, exc)
         if isinstance(exc, InfiniteRecursiveReference):
@@ -1405,7 +1401,7 @@ def find_invalid_headers(headers: Mapping) -> Generator[tuple[str, str], None, N
 
 UnsatisfiableExampleMark = Mark[Unsatisfiable](attr_name="unsatisfiable_example")
 NonSerializableMark = Mark[SerializationNotPossible](attr_name="non_serializable")
-InvalidRegexMark = Mark[SchemaError | ValidationError](attr_name="invalid_regex")
+InvalidRegexMark = Mark[ValidationError](attr_name="invalid_regex")
 InvalidHeadersExampleMark = Mark[dict[str, str]](attr_name="invalid_example_header")
 MissingPathParameters = Mark[InvalidSchema](attr_name="missing_path_parameters")
 InfiniteRecursiveReferenceMark = Mark[InfiniteRecursiveReference](attr_name="infinite_recursive_reference")